+++ services/core-service/src/apps/apps.module.ts
@@ -6,6 +6,7 @@ import { DiscoverModule } from "./discover/discover.module"
 import { ExpenseTrackModule } from "./expensetrack/expensetrack.module"
 import { TaxAdvisorModule } from "./taxadvisor/taxadvisor.module"
 import { CashFlowModule } from "./cashflow/cashflow.module"
+import { PlannerModule } from "./planner/planner.module"
 
 @Module({
   imports: [
@@ -16,6 +17,7 @@ import { CashFlowModule } from "./cashflow/cashflow.module"
     DiscoverModule,
     TaxAdvisorModule,
     CashFlowModule,
+    PlannerModule,
   ],
 })
 export class AppsModule {}