+++ services/core-service/src/apps/planner/event/commands/handler/create-event.handler.ts
@@ -1,0 +1,22 @@
+import { ICommandHandler, CommandHandler } from "@nestjs/cqrs"
+import { CreateEventCommand } from "../impl/create-event.command"
+import { EventRepository } from "../../event.repository"
+import { createOrConvertObjectId } from "@/shared/entity/entity.schema"
+import { toDateOnlyUTC } from "@/apps/cashflow/helpers/to-date"
+
+@CommandHandler(CreateEventCommand)
+export class CreateEventCommandHandler implements ICommandHandler<CreateEventCommand> {
+  constructor(private readonly repository: EventRepository) {}
+
+  async execute(command: CreateEventCommand) {
+    const {
+      userId,
+      dto: { eventDate, eventName },
+    } = command
+    return await this.repository.create({
+      userId: createOrConvertObjectId(userId),
+      eventName,
+      eventDate: toDateOnlyUTC(eventDate),
+    })
+  }
+}