+++ services/core-service/src/apps/planner/event/commands/impl/create-event.command.ts
@@ -1,0 +1,8 @@
+import { CreateEventRequestDto } from "../../dto/request/create-event.request.dto"
+
+export class CreateEventCommand {
+  constructor(
+    public readonly userId: string,
+    public readonly dto: CreateEventRequestDto
+  ) {}
+}