+++ services/core-service/src/apps/planner/event/dto/request/create-event.request.dto.ts
@@ -1,0 +1,9 @@
+import { IsDateString, IsNotEmpty } from "class-validator"
+
+export class CreateEventRequestDto {
+  @IsDateString()
+  eventDate: Date
+
+  @IsNotEmpty()
+  eventName: string
+}